+++ ffxivbot/pika_rabbit.py
@@ -37,301 +37,20 @@ import json
 from asgiref.sync import async_to_sync
 import ffxivbot.handlers as handlers
 from ffxivbot.models import *
+from ffxivbot.api_caller import ApiCaller
+from ffxivbot.event_handler import EventHandler
 
 
 USE_GRAFANA = getattr(settings, "USE_GRAFANA", False)
-
-
 CONFIG_PATH = os.environ.get(
     "FFXIVBOT_CONFIG", os.path.join(FFXIVBOT_ROOT, "ffxivbot/config.json")
 )
-
-
-# LOG_FORMAT = ('%(levelname) -10s %(asctime)s %(name) -30s %(funcName) '
-#               '-35s %(lineno) -5d: %(message)s')
 LOGGER = logging.getLogger(__name__)
 
 
-def handle_message(bot, message):
-    new_message = message
-    if isinstance(message, list):
-        new_message = []
-        for idx, msg in enumerate(message):
-            if msg["type"] == "share" and bot.share_banned:
-                share_data = msg["data"]
-                # new_message.append(
-                #     {
-                #         "type": "image",
-                #         "data": {
-                #             "file": share_data["image"],
-                #             "url": share_data["image"],
-                #         },
-                #     }
-                # )
-                new_message.append(
-                    {
-                        "type": "text",
-                        "data": {
-                            "text": "{}\n{}\n{}".format(
-                                share_data["title"],
-                                share_data["content"],
-                                share_data["url"],
-                            )
-                        },
-                    }
-                )
-            else:
-                new_message.append(msg)
-    return new_message
-
-
-def call_api(bot, action, params, echo=None, **kwargs):
-    if "async" not in action and not echo:
-        action = action + "_async"
-    if "send_" in action and "_msg" in action:
-        params["message"] = handle_message(bot, params["message"])
-    jdata = {"action": action, "params": params}
-    if echo:
-        jdata["echo"] = echo
-    post_type = kwargs.get("post_type", "websocket")
-    if post_type == "websocket":
-        async_to_sync(channel_layer.send)(
-            bot.api_channel_name, {"type": "send.event", "text": json.dumps(jdata)}
-        )
-    elif post_type == "http":
-        url = os.path.join(
-            bot.api_post_url, "{}?access_token={}".format(action, bot.access_token)
-        )
-        headers = {"Content-Type": "application/json"}
-        r = requests.post(url=url, headers=headers, data=json.dumps(params), timeout=5)
-        if r.status_code != 200:
-            print("HTTP Callback failed:")
-            print(r.text)
-    elif post_type == "wechat":
-        print("calling api:{}".format(action))
-
-        def req_url(params):
-            url = "https://ex-api.botorange.com/message/send"
-            headers = {"Content-Type": "application/json"}
-            print("params:{}".format(json.dumps(params)))
-            r = requests.post(
-                url=url, headers=headers, data=json.dumps(params), timeout=5
-            )
-            if r.status_code != 200:
-                print("Wechat HTTP Callback failed:")
-                print(r.text)
-
-        config = json.load(open(CONFIG_PATH, encoding="utf-8"))
-        params["chatId"] = kwargs.get("chatId", "")
-        params["token"] = config.get("WECHAT_TOKEN", "")
-        if "send_" in action and "_msg" in action:
-            if isinstance(params["message"], str):
-                text = params["message"]
-                at = re.finditer(r"\[CQ:at,qq=(.*)\]", text)
-                if at:
-                    params["mention"] = [at_m.group(1) for at_m in at]
-                text = re.sub(r"\[CQ:at,qq=(.*)\]", "", text)
-                img_r = r"\[CQ:image,file=(.*?)(?:\]|,.*?\])"
-                img_m = re.search(img_r, text)
-                if img_m:  # FIXME: handle text & img message
-                    params["messageType"] = 1
-                    params["payload"] = {"url": img_m.group(1)}
-                else:
-                    params["messageType"] = 0
-                    params["payload"] = {"text": text.strip()}
-                req_url(params)
-            else:
-                for msg_seg in params["message"]:
-                    if msg_seg["type"] == "image":
-                        params["messageType"] = 1
-                        params["payload"] = {"url": msg_seg["data"]["file"]}
-                        req_url(params)
-                    elif msg_seg["type"] == "text":
-                        params["messageType"] = 0
-                        params["payload"] = {"text": msg_seg["data"]["text"].strip()}
-                        req_url(params)
-                    time.sleep(1)
-
-    elif post_type == "tomon":
-        if "send_" in action and "_msg" in action:
-            print("Tomon Message >>> {}".format(params["message"]))
-            attachments = []
-            if isinstance(params["message"], str):
-                message = params["message"]
-                message = re.sub(r"\[CQ:at,qq=(.*?)\]", "<@\g<1>>", message)
-                print("message 1 >>> {}".format(message))
-                img_pattern = r"\[CQ:image,(?:cache=.,)?file=(.*?)\]"
-                m = re.search(img_pattern, message)
-                if m:
-                    attachments.append({"url": m.group(1)})
-                    message = re.sub(img_pattern, "", message)
-                    print("message 2 >>> {}".format(message))
-            elif isinstance(params["message"], list):
-                message = ""
-                for msg in params["message"]:
-                    if msg["type"] == "text":
-                        message += msg["data"]["text"]
-                    elif msg["type"] == "image":
-                        img_url = msg["data"]["file"]
-                        attachments.append({"url": img_url})
-                    elif msg["type"] == "share":
-                        share_data = msg["data"]
-                        message += "{}\n{}\n{}\n".format(
-                            share_data["title"],
-                            share_data["content"],
-                            share_data["url"],
-                        )
-            # if attachments:
-            #     for img in attachments:
-            #         message += img["url"] + " "
-            nonce = kwargs.get("nonce", "")
-            data = {"content": message, "nonce": nonce}
-            channel_id = kwargs.get("channel_id") or params.get("group_id")
-            url = "https://beta.tomon.co/api/v1/channels/{}/messages".format(channel_id)
-            headers = {
-                "Authorization": "Bearer {}".format(bot.tomon_bot.all()[0].token),
-            }
-            if attachments:
-                payload = {"payload_json": json.dumps(data)}
-                img_format = attachments[0]["url"].split(".")[-1]
-                original_image = requests.get(attachments[0]["url"], timeout=3)
-                files = [("image.{}".format(img_format), original_image.content)]
-                print("Posting Multipart to Tomon >>> {}".format(action))
-                print("{}".format(url))
-                r = requests.post(
-                    headers=headers, url=url, files=files, data=payload, timeout=30,
-                )
-                print(headers)
-                print(r.text)
-                if r.status_code != 200:
-                    print("Tomon HTTP Callback failed:")
-                    print(r.text)
-                return
-            headers.update({"Content-Type": "application/json"})
-            # print("Posting Json to Tomon >>> {}".format(action))
-            # print("{}".format(url))
-            # print("{}".format(json.dumps(data)))
-            r = requests.post(
-                url=url, headers=headers, data=json.dumps(data), timeout=3
-            )
-            if r.status_code != 200:
-                print("Tomon HTTP Callback failed:")
-                print(r.text)
-
-    elif post_type == "iotqq":
-        # print("IOTQQ action:")
-        # print(json.dumps(action, indent=4))
-        # print("IOTQQ params:")
-        # print(json.dumps(params, indent=4))
-        headers = {
-            "Content-Type": "application/json",
-        }
-        if bot.iotqq_auth:
-            headers.update(
-                {
-                    "Authorization": "Basic "
-                    + base64.b64encode(bot.iotqq_auth.encode()).decode()
-                }
-            )
-        if "send_" in action and "_msg" in action:
-            send_params = (
-                ("qq", bot.user_id),
-                ("funcname", "SendMsg"),
-            )
-            send_data = {
-                "toUser": params["group_id"],
-                "sendToType": 2,
-                "sendMsgType": "TextMsg",
-                "content": "",
-                "groupid": 0,
-                "atUser": 0,
-                "replayInfo": None,
-            }
-            message = params["message"]
-            attachments = []
-            if isinstance(params["message"], str):
-                message = re.sub(r"\[CQ:at,qq=(.*)\]", "[ATUSER(\g<1>)]", message)
-                img_pattern = r"\[CQ:image,(?:cache=.,)?file=(.*?)\]"
-                m = re.search(img_pattern, message)
-                if m:
-                    attachments.append({"url": m.group(1)})
-                    # message = re.sub(img_pattern, " \g<1> ", message)
-                    message = re.sub(img_pattern, "", message)
-            elif isinstance(params["message"], list):
-                message = ""
-                for msg in params["message"]:
-                    if msg["type"] == "text":
-                        message += msg["data"]["text"]
-                    elif msg["type"] == "image":
-                        img_url = msg["data"]["file"]
-                        attachments.append({"url": img_url})
-                    elif msg["type"] == "share":  # TODO: change to actual share
-                        share_data = msg["data"]
-                        message += "{}\n{}\n{}\n".format(
-                            share_data["title"],
-                            share_data["content"],
-                            share_data["url"],
-                        )
-            if attachments:
-                send_data["sendMsgType"] = "PicMsg"
-                send_data["picUrl"] = attachments[0]["url"]
-                send_data["picBase64Buf"] = ""
-                send_data["fileMd5"] = ""
-            if isinstance(message, str) and len(message) > 960:
-                message = message[:950] + "\n......"
-            send_data["content"] = message
-            print("IOTQQ send_data:")
-            print(json.dumps(send_data, indent=4))
-            r = requests.post(
-                bot.iotqq_url,
-                headers=headers,
-                params=send_params,
-                json=send_data,
-                timeout=10,
-            )
-            if r.status_code != 200:
-                print("IOTQQ HTTP Callback failed:")
-                print(r.text)
-
-    else:
-        LOGGER.error("Unsupported protocol: {}".format(post_type))
-
-
-def send_message(bot, private_group, uid, message, **kwargs):
-    if private_group == "group":
-        call_api(bot, "send_group_msg", {"group_id": uid, "message": message}, **kwargs)
-    if private_group == "discuss":
-        call_api(
-            bot, "send_discuss_msg", {"discuss_id": uid, "message": message}, **kwargs
-        )
-    if private_group == "private":
-        call_api(
-            bot, "send_private_msg", {"user_id": uid, "message": message}, **kwargs
-        )
-
-
-def update_group_member_list(bot, group_id, **kwargs):
-    call_api(
-        bot,
-        "get_group_member_list",
-        {"group_id": group_id},
-        "get_group_member_list:%s" % (group_id),
-        **kwargs,
-    )
-
-
-class PikaException(Exception):
-    """
-    Custom exception types
-    """
-
-    def __init__(self, message="Default PikaException"):
-        Exception.__init__(self, message)
-
-
 class PikaConsumer(object):
-    """This is based on the example consumer 
-    at https://github.com/pika/pika/blob/master/examples/asynchronous_consumer_example.py
+    """This is based on the example consumer at 
+    https://github.com/pika/pika/blob/master/examples/asynchronous_consumer_example.py
     """
 
     EXCHANGE = "message"
@@ -452,11 +171,12 @@ class PikaConsumer(object):
             receive["pika_time"] = time.time()
             self_id = receive["self_id"]
             try:
-                # get the bot
                 bot = QQBot.objects.get(user_id=self_id)
             except QQBot.DoesNotExist as e:
                 LOGGER.error("bot {} does not exsit.".format(self_id))
                 raise e
+            api_caller = ApiCaller(bot)
+            event_handler = EventHandler(bot, api_caller)
             config = json.load(open(CONFIG_PATH, encoding="utf-8"))
             already_reply = False
             if (
@@ -468,8 +188,7 @@ class PikaConsumer(object):
                         self_id, int(time.time())
                     )
                 )
-                call_api(
-                    bot,
+                api_caller.call_api(
                     "get_status",
                     {},
                     "get_status:{}".format(self_id),
@@ -477,417 +196,11 @@ class PikaConsumer(object):
                 )
 
             if receive["post_type"] == "message":
-                user_id = receive["user_id"]
-                # don't reply another bot
-                if QQBot.objects.filter(user_id=user_id).exists():
-                    raise PikaException(
-                        "{} reply from another bot:{}".format(
-                            receive["self_id"], user_id
-                        )
-                    )
-                (user, created) = QQUser.objects.get_or_create(user_id=user_id)
-                if 0 < time.time() < user.ban_till:
-                    raise PikaException(
-                        "User {} get banned till {}".format(user_id, user.ban_till)
-                    )
-
-                # replace alter commands
-                for (alter_command, command) in handlers.alter_commands.items():
-                    if receive["message"].find(alter_command) == 0:
-                        receive["message"] = receive["message"].replace(
-                            alter_command, command, 1
-                        )
-                        break
-
-                group_id = None
-                group = None
-                group_created = False
-                discuss_id = None
-                # Group Control Func
-                if receive["message"].find("\\") == 0:
-                    receive["message"] = receive["message"].replace("\\", "/", 1)
-                if receive["message_type"] == "discuss":
-                    discuss_id = receive["discuss_id"]
-                if receive["message_type"] == "group":
-                    group_id = receive["group_id"]
-                    (group, group_created) = QQGroup.objects.get_or_create(
-                        group_id=group_id
-                    )
-                    # self-ban in group
-                    if int(time.time()) < group.ban_till:
-                        raise PikaException(
-                            "{} banned by group:{}".format(self_id, group_id)
-                        )
-                        # LOGGER.info("{} banned by group:{}".format(self_id, group_id))
-                        # self.acknowledge_message(basic_deliver.delivery_tag)
-                        # return
-                    group_commands = json.loads(group.commands)
-                    group_bots = json.loads(group.bots)
-
-                    try:
-                        member_list = json.loads(group.member_list)
-                        if group_created or not member_list:
-                            update_group_member_list(
-                                bot,
-                                group_id,
-                                post_type=receive.get("reply_api_type", "websocket"),
-                            )
-                    except json.decoder.JSONDecodeError:
-                        member_list = []
-
-                    if receive["message"].find("/group_help") == 0:
-                        msg = (
-                            ""
-                            if member_list
-                            else "本群成员信息获取失败，请尝试重启酷Q并使用/update_group刷新群成员信息\n"
-                        )
-                        for (k, v) in handlers.group_commands.items():
-                            command_enable = True
-                            if group and group_commands:
-                                command_enable = (
-                                    group_commands.get(k, "enable") == "enable"
-                                )
-                            if command_enable:
-                                msg += "{}: {}\n".format(k, v)
-                        msg = msg.strip()
-                        send_message(
-                            bot,
-                            receive["message_type"],
-                            discuss_id or group_id or user_id,
-                            msg,
-                            post_type=receive.get("reply_api_type", "websocket"),
-                            chatId=receive.get("chatId", ""),
-                            channel_id=receive.get("channel_id", ""),
-                            nonce=receive.get("nonce", ""),
-                        )
-                    else:
-                        if receive["message"].find("/update_group") == 0:
-                            update_group_member_list(
-                                bot,
-                                group_id,
-                                post_type=receive.get("reply_api_type", "websocket"),
-                            )
-                        # get sender's user_info
-                        user_info = receive.get("sender")
-                        user_info = (
-                            user_info
-                            if (user_info and ("role" in user_info.keys()))
-                            else None
-                        )
-                        if member_list and not user_info:
-                            for item in member_list:
-                                if str(item["user_id"]) == str(user_id):
-                                    user_info = item
-                                    break
-                        if not user_info:
-                            if receive.get("reply_api_type", "websocket") == "wechat":
-                                user_info = {
-                                    "user_id": receive["user_id"],
-                                    "nickname": receive["data"]["contactName"],
-                                    "role": "member",
-                                }
-                                if receive["user_id"] not in list(
-                                    map(lambda x: x["user_id"], member_list)
-                                ):
-                                    member_list.append(user_info)
-                                    group.member_list = json.dumps(member_list)
-                                    group.save(update_fields=["member_list"])
-                            else:
-                                raise PikaException(
-                                    "No user info for user_id:{} in group:{}".format(
-                                        user_id, group_id
-                                    )
-                                )
-                            # LOGGER.error("No user info for user_id:{} in group:{}".format(user_id, group_id))
-                            # self.acknowledge_message(basic_deliver.delivery_tag)
-                            # return
-
-                        group_command_keys = sorted(
-                            handlers.group_commands.keys(), key=lambda x: -len(x)
-                        )
-                        for command_key in group_command_keys:
-                            if receive["message"].find(command_key) == 0:
-                                if (
-                                    receive["message_type"] == "group"
-                                    and group_commands
-                                ):
-                                    if (
-                                        command_key in group_commands.keys()
-                                        and group_commands[command_key] == "disable"
-                                    ):
-                                        continue
-                                if not group.registered and command_key != "/group":
-                                    msg = "本群%s未在数据库注册，请群主使用/register_group命令注册" % (
-                                        group_id
-                                    )
-                                    send_message(
-                                        bot,
-                                        "group",
-                                        group_id,
-                                        msg,
-                                        post_type=receive.get(
-                                            "reply_api_type", "websocket"
-                                        ),
-                                        chatId=receive.get("chatId", ""),
-                                        channel_id=receive.get("channel_id", ""),
-                                        nonce=receive.get("nonce", ""),
-                                    )
-                                    break
-                                else:
-                                    handle_method = getattr(
-                                        handlers,
-                                        "QQGroupCommand_{}".format(
-                                            command_key.replace("/", "", 1)
-                                        ),
-                                    )
-                                    action_list = handle_method(
-                                        receive=receive,
-                                        global_config=config,
-                                        bot=bot,
-                                        user_info=user_info,
-                                        member_list=member_list,
-                                        group=group,
-                                        commands=handlers.commands,
-                                        group_commands=handlers.group_commands,
-                                        alter_commands=handlers.alter_commands,
-                                    )
-                                    if USE_GRAFANA:
-                                        command_log = CommandLog(
-                                            time=int(time.time()),
-                                            bot_id=str(self_id),
-                                            user_id=str(user_id),
-                                            group_id=str(group_id),
-                                            command=str(command_key),
-                                            message=receive["message"],
-                                        )
-                                        command_log.save()
-                                    for action in action_list:
-                                        call_api(
-                                            bot,
-                                            action["action"],
-                                            action["params"],
-                                            echo=action["echo"],
-                                            post_type=receive.get(
-                                                "reply_api_type", "websocket"
-                                            ),
-                                            chatId=receive.get("chatId", ""),
-                                        )
-                                        already_reply = True
-                                    if already_reply:
-                                        break
-
-                if receive["message"].find("/help") == 0:
-                    msg = ""
-                    for (k, v) in handlers.commands.items():
-                        command_enable = True
-                        if group and group_commands:
-                            command_enable = group_commands.get(k, "enable") == "enable"
-                        if command_enable:
-                            msg += "{}: {}\n".format(k, v)
-                    msg += "具体介绍详见Wiki使用手册: {}\n".format(
-                        "https://github.com/Bluefissure/FFXIVBOT/wiki/"
-                    )
-                    msg = msg.strip()
-                    send_message(
-                        bot,
-                        receive["message_type"],
-                        group_id or user_id,
-                        msg,
-                        post_type=receive.get("reply_api_type", "websocket"),
-                        chatId=receive.get("chatId", ""),
-                        channel_id=receive.get("channel_id", ""),
-                        nonce=receive.get("nonce", ""),
-                    )
-
-                if receive["message"].find("/ping") == 0:
-                    time_from_receive = receive["time"]
-                    if time_from_receive > 3000000000:
-                        time_from_receive = time_from_receive / 1000
-                    msg = ""
-                    if "detail" in receive["message"]:
-                        msg += "[CQ:at,qq={}]\ncoolq->server: {:.2f}s\nserver->rabbitmq: {:.2f}s\nhandle init: {:.2f}s".format(
-                            receive["user_id"],
-                            receive["consumer_time"] - time_from_receive,
-                            receive["pika_time"] - receive["consumer_time"],
-                            time.time() - receive["pika_time"],
-                        )
-                    else:
-                        msg += "[CQ:at,qq={}] {:.2f}s".format(
-                            receive["user_id"], time.time() - time_from_receive
-                        )
-                    msg = msg.strip()
-                    LOGGER.info("{} calling command: {}".format(user_id, "/ping"))
-                    print(("{} calling command: {}".format(user_id, "/ping")))
-                    send_message(
-                        bot,
-                        receive["message_type"],
-                        discuss_id or group_id or user_id,
-                        msg,
-                        post_type=receive.get("reply_api_type", "websocket"),
-                        chatId=receive.get("chatId", ""),
-                        channel_id=receive.get("channel_id", ""),
-                        nonce=receive.get("nonce", ""),
-                    )
-
-                command_keys = sorted(handlers.commands.keys(), key=lambda x: -len(x))
-                for command_key in command_keys:
-                    if receive["message"].find(command_key) == 0:
-                        if receive["message_type"] == "group" and group_commands:
-                            if group_commands.get(command_key, "enable") == "disable":
-                                continue
-                            if group_bots and receive["self_id"] not in group_bots:
-                                continue
-                        handle_method = getattr(
-                            handlers,
-                            "QQCommand_{}".format(command_key.replace("/", "", 1)),
-                        )
-                        action_list = handle_method(
-                            receive=receive, global_config=config, bot=bot
-                        )
-                        if USE_GRAFANA:
-                            command_log = CommandLog(
-                                time=int(time.time()),
-                                bot_id=str(self_id),
-                                user_id=str(user_id),
-                                group_id="private"
-                                if receive["message_type"] != "group"
-                                else str(group_id),
-                                command=str(command_key),
-                                message=receive["message"],
-                            )
-                            command_log.save()
-                        for action in action_list:
-                            call_api(
-                                bot,
-                                action["action"],
-                                action["params"],
-                                echo=action["echo"],
-                                post_type=receive.get("reply_api_type", "websocket"),
-                                chatId=receive.get("chatId", ""),
-                                channel_id=receive.get("channel_id", ""),
-                                nonce=receive.get("nonce", ""),
-                            )
-                            already_reply = True
-                        break
-
-                # handling chat
-                if receive["message_type"] == "group":
-                    if not already_reply:
-                        action_list = handlers.QQGroupChat(
-                            receive=receive,
-                            global_config=config,
-                            bot=bot,
-                            user_info=user_info,
-                            member_list=member_list,
-                            group=group,
-                            commands=handlers.commands,
-                            alter_commands=handlers.alter_commands,
-                        )
-                        # need fix: disable chat logging for a while
-                        # if USE_GRAFANA:
-                        #     command_log = CommandLog(
-                        #         time = int(time.time()),
-                        #         bot_id = str(self_id),
-                        #         user_id = str(user_id),
-                        #         group_id = "private" if receive["message_type"] != "group" else str(group_id),
-                        #         command = "/chat",
-                        #         message = receive["message"]
-                        #     )
-                        #     command_log.save()
-                        for action in action_list:
-                            call_api(
-                                bot,
-                                action["action"],
-                                action["params"],
-                                echo=action["echo"],
-                                post_type=receive.get("reply_api_type", "websocket"),
-                                chatId=receive.get("chatId", ""),
-                                channel_id=receive.get("channel_id", ""),
-                                nonce=receive.get("nonce", ""),
-                            )
-
-            CONFIG_GROUP_ID = config["CONFIG_GROUP_ID"]
+                event_handler.on_message(receive, config=config)
             if receive["post_type"] == "request":
-                if receive["request_type"] == "friend":  # Add Friend
-                    qq = receive["user_id"]
-                    flag = receive["flag"]
-                    if bot.auto_accept_friend:
-                        reply_data = {"flag": flag, "approve": True}
-                        call_api(
-                            bot,
-                            "set_friend_add_request",
-                            reply_data,
-                            post_type=receive.get("reply_api_type", "websocket"),
-                        )
-                if (
-                    receive["request_type"] == "group"
-                    and receive["sub_type"] == "invite"
-                ):  # Invite Group
-                    flag = receive["flag"]
-                    if bot.auto_accept_invite:
-                        reply_data = {
-                            "flag": flag,
-                            "sub_type": "invite",
-                            "approve": True,
-                        }
-                        call_api(
-                            bot,
-                            "set_group_add_request",
-                            reply_data,
-                            post_type=receive.get("reply_api_type", "websocket"),
-                        )
-                if (
-                    receive["request_type"] == "group"
-                    and receive["sub_type"] == "add"
-                    and str(receive["group_id"]) == CONFIG_GROUP_ID
-                ):  # Add Group
-                    flag = receive["flag"]
-                    user_id = receive["user_id"]
-                    qs = QQBot.objects.filter(owner_id=user_id)
-                    if qs.count() > 0:
-                        reply_data = {"flag": flag, "sub_type": "add", "approve": True}
-                        call_api(
-                            bot,
-                            "set_group_add_request",
-                            reply_data,
-                            post_type=receive.get("reply_api_type", "websocket"),
-                        )
-                        reply_data = {
-                            "group_id": CONFIG_GROUP_ID,
-                            "user_id": user_id,
-                            "special_title": "饲养员",
-                        }
-                        call_api(
-                            bot,
-                            "set_group_special_title",
-                            reply_data,
-                            post_type=receive.get("reply_api_type", "websocket"),
-                        )
+                event_handler.on_request(receive, config=config)
             if receive["post_type"] == "notice":
-                if receive["notice_type"] == "group_increase":
-                    group_id = receive["group_id"]
-                    user_id = receive["user_id"]
-                    try:
-                        group = QQGroup.objects.get(group_id=group_id)
-                        msg = group.welcome_msg.strip()
-                        if msg != "":
-                            msg = "[CQ:at,qq=%s]" % (user_id) + msg
-                            send_message(
-                                bot,
-                                "group",
-                                group_id,
-                                msg,
-                                post_type=receive.get("reply_api_type", "websocket"),
-                                chatId=receive.get("chatId", ""),
-                                channel_id=receive.get("channel_id", ""),
-                                nonce=receive.get("nonce", ""),
-                            )
-                    except Exception as e:
-                        traceback.print_exc()
-            # print(" [x] Received %r" % body)
-        except PikaException as pe:
-            traceback.print_exc()
-            LOGGER.error(pe)
+                event_handler.on_notice(receive, config=config)
         except Exception as e:
             traceback.print_exc()
             LOGGER.error(e)
@@ -895,7 +208,7 @@ class PikaConsumer(object):
         self.acknowledge_message(basic_deliver.delivery_tag)
 
     def acknowledge_message(self, delivery_tag):
-        LOGGER.info("pid:%s Acknowledging message %s", os.getpid(), delivery_tag)
+        LOGGER.info("Pid:%s Acknowledging message %s", os.getpid(), delivery_tag)
         self._channel.basic_ack(delivery_tag)
 
     def stop_consuming(self):
@@ -928,7 +241,6 @@ class PikaConsumer(object):
 
 
 def main():
-    # logging.basicConfig(level=logging.INFO, format=LOG_FORMAT)
     logging.basicConfig(level=logging.INFO)
     pikapika = PikaConsumer("amqp://guest:guest@localhost:5672/?heartbeat=600")
     try:
